Add mdoc verification (Take 3) (#3)
This PR adds initial mdoc verification. There's a TODO around what other
data should come out verification, in a way that'd make it as
doc-agnostic as possible. Right now it seems "issuer identity" and
"wallet identity" are likeliest candidates, but each doc format (at
least comparing mdoc to SD-JWT) defines its own style of trust chain so
it won't always be "an array of X.509 certs".

OID4VP support targets **Draft 24 (27 January 2025)** of the spec:

https://openid.net/specs/openid-4-verifiable-presentations-1_0-24.html

Anyway not letting perfect be the enemy of good, I can get mdoc
verification in and define the next batch of work as issues.