Switch to "credential1" from "cred1" (#19)
Using "cred1" as the default name of the first credential in an OID4VP
request is something carried over from very early explorations in making
presentation requests. I'm switching to "credential1" instead because I
think it looks a little nicer. I'm still not entirely in love with it
but short of forcing developers to choose credential IDs it's _fine for
now._